+++ airbyte-ci/connectors/metadata_service/orchestrator/orchestrator/assets/github_assets.py
@@ -3,7 +3,10 @@ from dagster import Output, asset
 GROUP_NAME = "github"
 
 @asset(required_resource_keys={"github_connectors_directory"}, group_name=GROUP_NAME)
-def source_controlled_connectors(context):
+def github_connector_folders(context):
+    """
+    Return a list of all the folders in the github connectors directory.
+    """
     github_connectors_directory = context.resources.github_connectors_directory
 
     folder_names = [item.name for item in github_connectors_directory if item.type == 'dir']
